Gen_AI_Engineer_RAG/requirements.txt
sentence-transformers
faiss-cpu

streamlit
numpy
python-dotenv
